Support 21570 Add givenNames and familyNames to person - Add new fields to edit and view pages with tooltip - Organize type files to respective api folders
