+++ packages/dina-ui/page-tests/person/__tests__/edit.test.tsx
@@ -1,8 +1,8 @@
 import { OperationsResponse } from "common-ui";
-import { Organization } from "packages/dina-ui/types/objectstore-api/resources/Organization";
+import { Organization } from "packages/dina-ui/types/agent-api/resources/Organization";
 import PersonEditPage from "../../../pages/person/edit";
 import { mountWithAppContext } from "../../../test-util/mock-app-context";
-import { Person } from "../../../types/objectstore-api/resources/Person";
+import { Person } from "../../../types/agent-api/resources/Person";
 
 // Mock out the Link component, which normally fails when used outside of a Next app.
 jest.mock("next/link", () => ({ children }) => <div>{children}</div>);
