+++ packages/dina-ui/pages/collecting-event/view.tsx
@@ -12,7 +12,7 @@ import { ResourceIdentifierObject } from "jsonapi-typescript";
 import { KitsuResponse } from "kitsu";
 import { WithRouterProps } from "next/dist/client/with-router";
 import { withRouter } from "next/router";
-import { Person } from "packages/dina-ui/types/objectstore-api/resources/Person";
+import { Person } from "packages/dina-ui/types/agent-api/resources/Person";
 import { useContext, useState } from "react";
 import { Footer, GroupFieldView, Head, Nav } from "../../components";
 import { AttachmentList } from "../../components/object-store/attachment-list/AttachmentList";
