+++ packages/dina-ui/pages/collector-group/edit.tsx
@@ -12,7 +12,7 @@ import {
 import { ResourceSelectField } from "common-ui/lib";
 import { NextRouter, useRouter } from "next/router";
 import { CollectorGroup } from "../../types/collection-api/resources/CollectorGroup";
-import { Person } from "packages/dina-ui/types/objectstore-api/resources/Person";
+import { Person } from "packages/dina-ui/types/agent-api/resources/Person";
 import { useContext } from "react";
 import { Head, Nav } from "../../components";
 import { DinaMessage, useDinaIntl } from "../../intl/dina-ui-intl";
