+++ packages/dina-ui/pages/organization/view.tsx
@@ -11,7 +11,7 @@ import { WithRouterProps } from "next/dist/client/with-router";
 import { withRouter } from "next/router";
 import { Footer, Head, Nav } from "../../components";
 import { DinaMessage, useDinaIntl } from "../../intl/dina-ui-intl";
-import { Organization } from "../../types/objectstore-api/resources/Organization";
+import { Organization } from "../../types/agent-api/resources/Organization";
 
 export function OrganizationDetailsPage({ router }: WithRouterProps) {
   const { id } = router.query;
