+++ packages/dina-ui/pages/person/edit.tsx
@@ -2,7 +2,7 @@ import { ButtonBar, CancelButton, LoadingSpinner, Query } from "common-ui";
 import { useRouter } from "next/router";
 import { Footer, Head, Nav, PersonForm } from "../../components";
 import { DinaMessage, useDinaIntl } from "../../intl/dina-ui-intl";
-import { Person } from "../../types/objectstore-api/resources/Person";
+import { Person } from "../../types/agent-api/resources/Person";
 
 export default function PersonEditPage() {
   const router = useRouter();
