+++ packages/dina-ui/types/objectstore-api/index.ts
@@ -1,7 +1,7 @@
 export * from "./resources/AuditSnapshot";
 export * from "./resources/DefaultValue";
 export * from "./resources/License";
-export * from "./resources/Person";
+export * from "../agent-api/resources/Person";
 export * from "./resources/ManagedAttribute";
 export * from "./resources/ManagedAttributeMap";
 export * from "./resources/Metadata";
